Make the part header be a layout class
